+++ src/components/domain/lead/details/Details.tsx
