+++ src/components/domain/lead/summary/Summary.tsx
